third_party/WebKit/LayoutTests/media/track/opera/track/track-element/cloneNode-expected.txt

This is a testharness.js-based test.
FAIL track element cloneNode, not loaded assert_equals: elm.readyState after setting mode expected 1 but got 0
FAIL track element cloneNode, loading assert_equals: elm.readyState after setting mode expected 1 but got 0
FAIL track element cloneNode, loaded assert_equals: elm.readyState after setting mode expected 1 but got 0
Harness: the test ran to completion.
